requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
mistralai==0.4.2
python-dotenv==1.0.1
gunicorn==21.2.0
flask
requests
streamlit
PyPDF2
pdfplumber
pdf2image
pytesseract
gdown
langchain
nltk
transformers
torch
sentence-transformers
